feat(backend): add feature update agent avatar

+++ chatbot-core/backend/app/services/agent.py
@@ -1,12 +1,8 @@
-import hashlib
-import os
-from datetime import datetime
 from io import BytesIO
 from typing import List
 from typing import Optional
 from typing import Tuple
 
-import pydenticon
 from fastapi import UploadFile
 from sqlalchemy.orm import Session
 
@@ -18,8 +14,9 @@ from app.services.base import BaseService
 from app.settings import Constants
 from app.utils.api.api_response import APIError
 from app.utils.api.error_handler import ErrorCodesMappingNumber
+from app.utils.api.helpers import construct_file_path
+from app.utils.api.helpers import generate_avatar_image
 from app.utils.api.helpers import get_logger
-from app.utils.api.helpers import remove_vietnamese_accents
 
 logger = get_logger(__name__)
 
@@ -63,6 +60,53 @@ class AgentService(BaseService):
         """
         return self._agent_repo.get_agent(agent_id=agent_id, user_id=user_id)
 
+    def _upload_agent_avatar(
+        self,
+        agent_avatar: BytesIO,
+        agent_name: str,
+        user_id: str,
+        delete_existing_image: bool = False,
+        existing_image_path: str = None,
+    ) -> Tuple[str, Optional[APIError]]:
+        """
+        Upload agent avatar to Minio.
+
+        Args:
+            agent_avatar (BytesIO): Agent avatar image
+            agent_name (str): Agent name
+            user_id (str): User id
+            delete_existing_image (bool, optional): Whether to delete existing image. Defaults to False.
+            existing_image_path (str, optional): Existing image path. Defaults to None.
+
+        Returns:
+            Tuple[str, Optional[APIError]]: File path in Minio and APIError object if any error
+        """
+        # Construct file path in Minio
+        file_path = construct_file_path(object_name=agent_name, user_id=user_id)
+        logger.info(f"Uploading agent avatar to Minio: {file_path}")
+
+        # Upload image to Minio
+        is_file_uploaded = self._minio_connector.upload_file(
+            object_name=file_path,
+            data=agent_avatar,
+            bucket_name=Constants.MINIO_IMAGE_BUCKET,
+        )
+        if not is_file_uploaded:
+            return "", APIError(kind=ErrorCodesMappingNumber.UNABLE_TO_UPLOAD_FILE_TO_MINIO.value)
+
+        # Delete existing image if needed
+        if delete_existing_image and existing_image_path:
+            logger.info(f"Deleting existing agent avatar in Minio: {existing_image_path}")
+            is_file_deleted = self._minio_connector.delete_file(
+                object_name=existing_image_path, bucket_name=Constants.MINIO_IMAGE_BUCKET
+            )
+            if not is_file_deleted:
+                return "", APIError(
+                    kind=ErrorCodesMappingNumber.UNABLE_TO_DELETE_FILE_IN_MINIO.value
+                )
+
+        return file_path, None
+
     def create_agent(
         self, agent_request: AgentRequest, user_id: str, file: Optional[UploadFile] = None
     ) -> Optional[APIError]:
@@ -78,50 +122,19 @@ class AgentService(BaseService):
             Optional[APIError]: APIError object if any error
         """
         with self._transaction():
-            # Construct file path in Minio
-            file_name = (
-                (remove_vietnamese_accents(input_str=agent_request.name).replace(" ", "_").lower())
-                + "_"
-                + user_id
-                + datetime.now().strftime("%Y%m%d%H%M%S")
-                + ".png"
-            )
-            logger.info(f"File name: {file_name}")
-            file_path = os.path.join(Constants.MINIO_IMAGE_BUCKET, file_name)
-
             # Upload image to Minio if file is provided. Otherwise, use the generated image.
             agent_avatar = None
             if not file:
-                generator = pydenticon.Generator(
-                    5,
-                    5,
-                    digest=hashlib.sha1,
-                    foreground=Constants.AGENT_AVATAR_IDENTICON_FOREGROUND_COLOR,
-                    background=Constants.AGENT_AVATAR_IDENTICON_BACKGROUND_COLOR,
-                )
-
-                # Generate the image
-                image_data = generator.generate(
-                    data=file_name,
-                    width=Constants.AGENT_AVATAR_IDENTICON_WIDTH,
-                    height=Constants.AGENT_AVATAR_IDENTICON_HEIGHT,
-                    output_format=Constants.AGENT_AVATAR_IDENTICON_OUTPUT_FORMAT,
-                )
-
-                # Convert to BinaryIO
-                agent_avatar = BytesIO(image_data)
+                agent_avatar = generate_avatar_image(data=agent_request.name)
             else:
                 agent_avatar = file.file
 
             # Upload image to Minio
-            logger.info(f"Uploading agent avatar to Minio: {file_path}")
-            is_file_uploaded = self._minio_connector.upload_files(
-                object_name=file_path,
-                data=agent_avatar,
-                bucket_name=Constants.MINIO_IMAGE_BUCKET,
+            file_path, err = self._upload_agent_avatar(
+                agent_avatar=agent_avatar, agent_name=agent_request.name, user_id=user_id
             )
-            if not is_file_uploaded:
-                return APIError(kind=ErrorCodesMappingNumber.UNABLE_TO_UPLOAD_FILE_TO_MINIO.value)
+            if err:
+                return err
 
             # Define agent
             agent = Agent(
@@ -138,7 +151,11 @@ class AgentService(BaseService):
         return err if err else None
 
     def update_agent(
-        self, agent_id: str, agent_request: AgentRequest, user_id: str
+        self,
+        agent_id: str,
+        agent_request: AgentRequest,
+        user_id: str,
+        file: Optional[UploadFile] = None,
     ) -> Optional[APIError]:
         """
         Update an agent.
@@ -147,13 +164,35 @@ class AgentService(BaseService):
             agent_id (str): Agent id
             agent_request (AgentRequest): Agent request object
             user_id (str): User id
+            file (Optional[UploadFile]): File object. Defaults to None.
 
         Returns:
             Optional[APIError]: APIError object if any error
         """
         with self._transaction():
             # Define to-be-updated agent
-            agent = agent_request.model_dump(exclude_unset=True)
+            agent = agent_request.model_dump(exclude_unset=True, exclude_defaults=True)
+
+            # Upload new image to Minio if file is provided.
+            if file:
+                # Get existing agent
+                existing_agent, err = self._agent_repo.get_agent(agent_id=agent_id, user_id=user_id)
+                if err:
+                    return err
+
+                # Upload image to Minio
+                file_path, err = self._upload_agent_avatar(
+                    agent_avatar=file.file,
+                    agent_name=existing_agent.name,
+                    user_id=user_id,
+                    delete_existing_image=True,
+                    existing_image_path=existing_agent.uploaded_image_path,
+                )
+                if err:
+                    return err
+
+                # Update agent with new image
+                agent["uploaded_image_path"] = file_path
 
             # Update agent
             err = self._agent_repo.update_agent(agent_id=agent_id, agent=agent, user_id=user_id)
@@ -172,6 +211,23 @@ class AgentService(BaseService):
             Optional[APIError]: APIError object if any error
         """
         with self._transaction():
+            # Get existing agent
+            existing_agent, err = self._agent_repo.get_agent(agent_id=agent_id, user_id=user_id)
+            if err:
+                return err
+
+            if existing_agent.uploaded_image_path:
+                # Delete image in Minio
+                logger.info(f"Deleting agent avatar in Minio: {existing_agent.uploaded_image_path}")
+                is_file_deleted = self._minio_connector.delete_file(
+                    object_name=existing_agent.uploaded_image_path,
+                    bucket_name=Constants.MINIO_IMAGE_BUCKET,
+                )
+                if not is_file_deleted:
+                    return APIError(
+                        kind=ErrorCodesMappingNumber.UNABLE_TO_DELETE_FILE_IN_MINIO.value
+                    )
+
             # Delete agent
             err = self._agent_repo.delete_agent(agent_id=agent_id, user_id=user_id)
